move keyManager to relayer package
